renamed Magma Spawn to Infernal
