+++ assets/units/data.json
@@ -333,7 +333,7 @@
 	},
 	{
 		"id": 4,
-		"name": "Magma Spawn",
+		"name": "Infernal",
 		"level": 2,
 		"realm": "L",
 		"size": 3,
