+++ assets/units/sprites/Infernal - Scorched Ground.png
