+++ src/abilities/Infernal.js
