+++ src/index.ejs
@@ -652,9 +652,9 @@
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Magma_Spawn" class="vignette realmL typeL2" creature="L2">
+						<a href="#Infernal" class="vignette realmL typeL2" creature="L2">
 							<div class="tooltip">
-								<div class="content">Magma Spawn</div>
+								<div class="content">Infernal</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
