+++ src/style/avatars.less
@@ -88,7 +88,7 @@
 	}
 
 	&.typeL2 {
-		background-image: url('~assets/units/avatars/Magma Spawn.jpg');
+		background-image: url('~assets/units/avatars/Infernal.jpg');
 	}
 
 	&.typeL3 {
