+++ src/style/styles.less
@@ -1313,11 +1313,11 @@ span.pure {
 }
 
 .trap.scorched-ground {
-	background: url('~assets/units/sprites/Magma Spawn - Scorched Ground.png');
+	background: url('~assets/units/sprites/Infernal - Scorched Ground.png');
 }
 
 .trap.firewall {
-	background: url('~assets/units/sprites/Magma Spawn - Scorched Ground.png');
+	background: url('~assets/units/sprites/Infernal - Scorched Ground.png');
 }
 
 /*------------------Progress Bars-------------------*/
